+++ cy-news-server/cy-news-common/src/main/java/com/cy/news/common/Pojo/CommentsMessage.java
@@ -18,7 +18,7 @@ import java.io.Serializable;
 public class CommentsMessage implements Serializable {
     private Integer mId;
 
-    private Integer cId;
+    private Long cId;
 
     private Integer mLike;
