+++ cy-news-server/cy-news-common/src/main/java/com/cy/news/common/Pojo/NewsMessage.java
@@ -18,7 +18,7 @@ import java.io.Serializable;
 public class NewsMessage implements Serializable {
     private Integer mId;
 
-    private Integer nId;
+    private Long nId;
 
     private Integer mLike;
